backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "emanuelle",email = "emanuellefelix.dev@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.1.5,<6.0.0)",
    "psycopg (>=3.2.4,<4.0.0)",
    "psycopg-binary (>=3.2.4,<4.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "django-ninja (>=1.3.0,<2.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
